Ver 0.2 Statisztika rögzítési felület elkészítése

+++ app/src/main/java/hu/szakdolgozat/handballstatistics/models/Player.java
@@ -3,18 +3,18 @@ package hu.szakdolgozat.handballstatistics.models;
 import androidx.annotation.NonNull;
 
 public class Player {
-    private final int playerId;
+    private final long playerId;
     private final String name;
     private final String team;
 
 
-    public Player(int playerId, String name, String team) {
+    public Player(long playerId, String name, String team) {
         this.playerId = playerId;
         this.name = name;
         this.team = team;
     }
 
-    public int getId() {
+    public long getId() {
         return playerId;
     }
